Add docs `Ul::class`.
